Original change: https://googleplex-android-review.googlesource.com/c/platform/packages/modules/HealthFitness/+/21642782

+++ framework/java/android/health/connect/datatypes/HeartRateRecord.java
@@ -308,7 +308,7 @@ public final class HeartRateRecord extends IntervalRecord {
         for (HeartRateRecord.HeartRateSample heartRateSample : getSamples()) {
             samples.add(
                     new HeartRateRecordInternal.HeartRateSample(
-                            heartRateSample.getBeatsPerMinute(),
+                            (int) heartRateSample.getBeatsPerMinute(),
                             heartRateSample.getTime().toEpochMilli()));
         }
         recordInternal.setSamples(samples);
